Mini Project 2
Objective
The purpose of this assignment is to explore Java EE service layer
components of an enterprise application, integrating JPA domain model
with the service layer, and introducing security to the application.

About the application:
Global Parking System is a parking application which keeps information
about the vehicle, parking details, Billing information that are parked
in XYZ Company’s paid parking arena. In this paid parking arena,
customers are supposed to pay for parking their vehicle on hourly
manner.
This application will have two actors, employee and customer. Where each
actor will have predefined role and functionality
In this assignment to showcase security implementation there are
different behavior for each role,  I.e, Customer can view only his
information alone and employee and see all the customer’s and employee
details.

MiniProject4.docx is the read me file. , This project is still in
progress and need to implement JSF s and JavaEE concepts to make it full
fledge Parking Application.